+++ src/server/components/dashboard/relatedLinks/handlers/index.ts
@@ -1,0 +1,2 @@
+export * as edit from "./edit";
+export * as confirm from "./confirm";
