+++ src/server/components/dashboard/relatedLinks/router.ts
@@ -4,6 +4,7 @@ import { logger } from "server/services/logger";
 import { relatedLinkSchema } from "server/components/dashboard/relatedLinks/relatedLinkSchema";
 import Joi from "joi";
 import { RelatedLink } from "shared/types";
+import { confirm, edit } from "./handlers";
 
 export const relatedLinksRouter = express.Router();
 
@@ -45,94 +46,10 @@ relatedLinksRouter.get("/", (req, res, next) => {
   res.redirect(res.locals.listsEditUrl);
 });
 
-relatedLinksRouter.get("/:relatedLinkIndex", (req, res, next) => {
-  const relatedLinkErrorSummary = req.flash("relatedLinkErrorSummary") as unknown as string[];
-  res.render("dashboard/related-links/edit", {
-    relatedLinkErrorSummary: relatedLinkErrorSummary.map((error) => JSON.parse(error)),
-  });
-});
+relatedLinksRouter.get("/:relatedLinkIndex", edit.get);
 
-relatedLinksRouter.post("/:relatedLinkIndex", (req, res, next) => {
-  const { text, url } = req.body;
-  const { relatedLinkIndex } = req.params;
+relatedLinksRouter.post("/:relatedLinkIndex", edit.post);
 
-  req.session.relatedLink = {
-    text,
-    url,
-  };
+relatedLinksRouter.get("/:relatedLinkIndex/confirm", confirm.get);
 
-  const { error } = relatedLinkSchema.validate(
-    { text, url },
-    {
-      abortEarly: false,
-      errors: {
-        wrap: {
-          label: false,
-        },
-      },
-    }
-  );
-
-  if (error) {
-    error.details.forEach((detail) => {
-      const { key } = detail.context!;
-      const error = {
-        text: detail.message,
-        href: key,
-      };
-      req.flash("relatedLinkErrorSummary", JSON.stringify(error));
-      req.flash(`relatedLinkError_${key}`, detail.message);
-    });
-
-    req.session.relatedLink = {
-      text,
-      url,
-    };
-
-    return res.redirect(`${relatedLinkIndex}`);
-  }
-
-  res.redirect(`${relatedLinkIndex}/confirm`);
-});
-
-relatedLinksRouter.get("/:relatedLinkIndex/confirm", (req, res, next) => {
-  const { relatedLinkIndex } = req.params;
-  const { text, url } = req.session.relatedLink ?? {};
-
-  if (!text || !url) {
-    return res.redirect(`${res.locals.listsEditUrl}/related-links/${relatedLinkIndex}`);
-  }
-
-  res.render("dashboard/related-links/confirm", {});
-});
-
-relatedLinksRouter.post("/:relatedLinkIndex/confirm", async (req, res, next) => {
-  const { id } = res.locals.list;
-  const { relatedLinkIndex } = req.params;
-  const { text, url } = req.session.relatedLink ?? {};
-
-  if (!text || !url) {
-    await res.redirect(`/related-links/${relatedLinkIndex}`);
-  }
-
-  const update: RelatedLink = {
-    text: text!,
-    url: url!,
-  };
-
-  try {
-    const transaction = await addRelatedLink(id, update);
-    if (transaction) {
-      req.flash("successBannerHeading", "Success");
-      req.flash("successBannerMessage", `The link ${text} was added successfully`);
-    }
-  } catch (e) {
-    logger.error(`User ${req.user?.id} attempted to update ${req.originalUrl} failed with ${e}`);
-
-    req.flash("error", `Adding the link ${text} failed`);
-  }
-
-  delete req.session.relatedLink;
-
-  await res.redirect(`${res.locals.listsEditUrl}`);
-});
+relatedLinksRouter.post("/:relatedLinkIndex/confirm", confirm.post);
